+++ Classes/Model/Type/Book.php
@@ -20,7 +20,7 @@ use Brotkrueml\Schema\Manual\Publisher;
  * A book.
  */
 #[Type('Book')]
-#[Manual(Publisher::Google, 'https://developers.google.com/search/docs/data-types/book')]
+#[Manual(Publisher::Google, 'https://developers.google.com/search/docs/appearance/structured-data/book')]
 final class Book extends AbstractType
 {
     protected static array $propertyNames = [