+++ Classes/Model/Type/BreadcrumbList.php
@@ -22,7 +22,7 @@ use Brotkrueml\Schema\Manual\Publisher;
  * The position property is used to reconstruct the order of the items in a BreadcrumbList. The convention is that a breadcrumb list has an itemListOrder of ItemListOrderAscending (lower values listed first), and that the first items in this list correspond to the "top" or beginning of the breadcrumb trail, e.g. with a site or section homepage. The specific values of 'position' are not assigned meaning for a BreadcrumbList, but they should be integers, e.g. beginning with '1' for the first item in the list.
  */
 #[Type('BreadcrumbList')]
-#[Manual(Publisher::Google, 'https://developers.google.com/search/docs/data-types/breadcrumb')]
+#[Manual(Publisher::Google, 'https://developers.google.com/search/docs/appearance/structured-data/breadcrumb')]
 final class BreadcrumbList extends AbstractType
 {
     protected static array $propertyNames = [