+++ Classes/Model/Type/ContactPoint.php
@@ -11,13 +11,16 @@ declare(strict_types=1);
 
 namespace Brotkrueml\Schema\Model\Type;
 
+use Brotkrueml\Schema\Attributes\Manual;
 use Brotkrueml\Schema\Attributes\Type;
 use Brotkrueml\Schema\Core\Model\AbstractType;
+use Brotkrueml\Schema\Manual\Publisher;
 
 /**
  * A contact point - for example, a Customer Complaints department.
  */
 #[Type('ContactPoint')]
+#[Manual(Publisher::Google, 'https://developers.google.com/search/docs/appearance/structured-data/logo')]
 final class ContactPoint extends AbstractType
 {
     protected static array $propertyNames = [