+++ Classes/Model/Type/DiscussionForumPosting.php
@@ -11,13 +11,16 @@ declare(strict_types=1);
 
 namespace Brotkrueml\Schema\Model\Type;
 
+use Brotkrueml\Schema\Attributes\Manual;
 use Brotkrueml\Schema\Attributes\Type;
 use Brotkrueml\Schema\Core\Model\AbstractType;
+use Brotkrueml\Schema\Manual\Publisher;
 
 /**
  * A posting to a discussion forum.
  */
 #[Type('DiscussionForumPosting')]
+#[Manual(Publisher::Google, 'https://developers.google.com/search/docs/appearance/structured-data/discussion-forum')]
 final class DiscussionForumPosting extends AbstractType
 {
     protected static array $propertyNames = [