+++ Classes/Model/Type/HomeAndConstructionBusiness.php
@@ -17,6 +17,8 @@ use Brotkrueml\Schema\Core\Model\AbstractType;
 /**
  * A construction business.
  *
+ * A HomeAndConstructionBusiness is a LocalBusiness that provides services around homes and buildings.
+ *
  * As a LocalBusiness it can be described as a provider of one or more Service\(s).
  */
 #[Type('HomeAndConstructionBusiness')]