+++ Classes/Model/Type/Movie.php
@@ -20,7 +20,7 @@ use Brotkrueml\Schema\Manual\Publisher;
  * A movie.
  */
 #[Type('Movie')]
-#[Manual(Publisher::Google, 'https://developers.google.com/search/docs/data-types/movie')]
+#[Manual(Publisher::Google, 'https://developers.google.com/search/docs/appearance/structured-data/movie')]
 #[Manual(Publisher::Yandex, 'https://yandex.com/support/webmaster/supported-schemas/movie-description.html')]
 final class Movie extends AbstractType
 {