+++ Classes/Model/Type/PostalAddress.php
@@ -11,13 +11,17 @@ declare(strict_types=1);
 
 namespace Brotkrueml\Schema\Model\Type;
 
+use Brotkrueml\Schema\Attributes\Manual;
 use Brotkrueml\Schema\Attributes\Type;
 use Brotkrueml\Schema\Core\Model\AbstractType;
+use Brotkrueml\Schema\Manual\Publisher;
 
 /**
  * The mailing address.
  */
 #[Type('PostalAddress')]
+#[Manual(Publisher::Google, 'https://developers.google.com/search/docs/appearance/structured-data/logo')]
+#[Manual(Publisher::Yandex, 'https://yandex.com/support/webmaster/supported-schemas/address-organization.html')]
 final class PostalAddress extends AbstractType
 {
     protected static array $propertyNames = [