+++ Classes/Model/Type/Product.php
@@ -20,7 +20,7 @@ use Brotkrueml\Schema\Manual\Publisher;
  * Any offered product or service. For example: a pair of shoes; a concert ticket; the rental of a car; a haircut; or an episode of a TV show streamed online.
  */
 #[Type('Product')]
-#[Manual(Publisher::Google, 'https://developers.google.com/search/docs/data-types/product')]
+#[Manual(Publisher::Google, 'https://developers.google.com/search/docs/appearance/structured-data/product')]
 #[Manual(Publisher::Yandex, 'https://yandex.com/support/webmaster/supported-schemas/goods-prices.html')]
 final class Product extends AbstractType
 {