+++ Classes/Model/Type/ProfilePage.php
@@ -11,14 +11,17 @@ declare(strict_types=1);
 
 namespace Brotkrueml\Schema\Model\Type;
 
+use Brotkrueml\Schema\Attributes\Manual;
 use Brotkrueml\Schema\Attributes\Type;
 use Brotkrueml\Schema\Core\Model\AbstractType;
 use Brotkrueml\Schema\Core\Model\WebPageTypeInterface;
+use Brotkrueml\Schema\Manual\Publisher;
 
 /**
  * Web page type: Profile page.
  */
 #[Type('ProfilePage')]
+#[Manual(Publisher::Google, 'https://developers.google.com/search/docs/appearance/structured-data/profile-page')]
 final class ProfilePage extends AbstractType implements WebPageTypeInterface
 {
     protected static array $propertyNames = [