+++ Classes/Model/Type/Recipe.php
@@ -20,7 +20,7 @@ use Brotkrueml\Schema\Manual\Publisher;
  * A recipe. For dietary restrictions covered by the recipe, a few common restrictions are enumerated via suitableForDiet. The keywords property can also be used to add more detail.
  */
 #[Type('Recipe')]
-#[Manual(Publisher::Google, 'https://developers.google.com/search/docs/data-types/recipe')]
+#[Manual(Publisher::Google, 'https://developers.google.com/search/docs/appearance/structured-data/recipe')]
 #[Manual(Publisher::Yandex, 'https://yandex.com/support/webmaster/supported-schemas/recipe.html')]
 final class Recipe extends AbstractType
 {